add mesh tests
